Be extra paranoid in annotate_state_set_scope
Previously it was possible to use the same state object to store a
per-message annotation, then change scope and use it again to store a
per-server annotation.

This worked because, although the data are stored in two different DBs,
the transaction model was global across DBs.   Now the transactional
model is one txn in one DB per state object, so we can't do that
anymore.

The actual imapd code never did this, but some of the unit tests did.
So, be slightly more paranoid and just assert that it's not happening.