Fix annotate_msg_copy() back up again
There's a subtle requirement here - that the destination mailbox must
have had mailbox_get_annotate_state() called on it, otherwise the data
copied will never get committed.  Also, we need to explicitly start a
txn.

Do that extra call in both the callers, and assert that it's true in
annotate_msg_copy().