cmulocal/berkdb.m4 Bug#3708 check for db_create instead of db_open
https://bugzilla.cyrusimap.org/show_bug.cgi?id=3708
From: Carson Gaspar

db_open is used nowhere in the code, and hasn't been part of Berkeley DB
since 2.x. Checking for db_create (the function actually used in the code)
is the correct test.

To verify the bug, build Berkeley DB 4.x or later (probably 3.x, but I
don't have anything that old to test against) with versioned symbols.
Watch configure fail (as the symbol name is db_open_${version}, and
db.h doesn't map it as it's a private interface)