config.h: add #define EXPORTED and HIDDEN
With gcc >= 4.0 symbols can be marked as hidden, so that they are not
included in the dynamic symbol table of exported symbols.

With the additions to configure.ac , a library compiled with CFLAGS =
$(CFLAG_VISIBILTY) will have by default all symbols hidden, and each can
be "shown" by prefixing it with the EXPORTE macro.  On the other side, in
the rest of the libraries are all symbols are by default exported, so each
symbol can be prefixed there with HIDDEN in order to exclude just that
symbol from the dynamic table.