Makefile.am: link with the right libcom_err
The used libcom_err (the system wide, or the bundled
com_err/et/libcyrus_com_err) were linked in the final executables with the
LD_UTILITY flag.  this lead to a sitation, where the libcyrus_sieve and
libcyrus_imap did not knew, on which COM_ERR library they depend on.

With this fix, libcyrus_imap and libcyrus_sieve learn to link explicitly
with the "right" libcom_err, and programs linking with libcyrus_imap
or libcyrus_sieve do not have to care on this: they just link with
libcyrus_imap.