libcyrus_imap: hide even more local symbols
Annotate with static all functions, that are defined (and possibly used)
in one and the same source file, and are separately declared in a
header file.

The idea is to be able to find the functions, that are intended to be used
by other source files, but are not, and could be later completely removed.