sync: automatically upgrade persistent CRCs
when calculating the CRC for a writable mailbox via the slowpath, for a
newer CRC version, save the resulting CRC and version in the mailbox so
we hit the fastpath next time.

Also, move the calculation function into mailbox.c.  Now sync_support.c
CRC code is a tiny wrapper around mailbox_*() functions.

Thanks to Bron for pointing out these optimisation in code review.

One unexpected side effect of doing the CRC slowpath in mailbox.c is
that we cannot just use the read_annotations() function, and instead
have to call annotatmore_findall() directly, which avoids quite a lot of
memory traffic because we now CRC the annotations as they're iterated
instead of saving them to memory structures and iterating those.