smmapd: just use the quota DB directly to find the quotaroot
This saves the heavy work of locking, loading, mapping in the
cyrus.index file only to use it to find the quotaroot anyway.

An alternative path would be to find the real cyrus.header file
and then load that to look for the quotaroot... but in practice
a couple of DB lookups on the quota DB are probably more efficent.