squatter: remove initial 32ms delay when rolling
The purpose of this delay was to give the lmtpd or imapd the chance to
drop the mboxname lock before we took it to index the mailbox.  This
turned out not to be a useful tactic because imapds can keep the
mboxname lock held for longer than squatter is willing to keep retrying.
As we are now taking the lock in shared mode for indexing, there's no
need for the initial 32ms delay before the first try.