[IRIS-2038] Report correct snippets
when the query is nontrivial, e.g. contains booleans or field modifiers.

It turns out that Sphinx' CALL SNIPPETS command doesn't work at all like
I thought it did.  It's actually completely unaware of fields and when
given field modifiers in the query it just happily ignores them.  So we
need to handle fields ourselves.  Also, we need to modify the query sent
to Sphinx to be just a single big disjunction of all the search terms in
the original query regardless of their actual position.