Bug #3715 - cyrus 2.4 only fix
This bug doesn't exist in either 2.3 or 2.5, for different reasons.

In 2.3, we actually lock the mailboxes.db while we do the create.

In 2.5, we don't do the RMW flag at all.

Ideally we would hold a namelock longer during CREATE to avoid the
spurious error message late, but this is still robust.