Revert "sieve/bc_eval: fix memory leak"
This reverts commit 55ac1abac8e4bc8c65351c4eebb4fda306937a8a.

Actually, the memory WAS freed - because it got passed to do_vacation
which, which does horrible things with putting it in a struct and then
freeing it when the struct finishes.

The whole lifetimes of allocated memory and ownership chain is horribly
complex and broken in a lot of sieve - but the way to fix it sure isn't
just blindy deallocating.  It's going to take a lot more work than this.