index: notice UIDVALIDITY changes and treat as deletes
This is necessary because without the namelock, the folder
may be deleted and then created again without our session
noticing, meaning the UIDs will be all wrong.

We can just treat the change the same way we would treat
a delete at the index level.  The user will be told that
the current mailbox has been deleted, and if they select
again they will see the new UIDVALIDITY.