imapd: actually release index between uses
The release is placed just before checking idle, and just before
reading from the command line in cmdloop.  This ensures that any
current action has a chance to finish reading from files or cache
without the repack racing against it, but doesn't need to have
changes made in hundreds of super-accurate lifetime-based places
in the code.