conversations: store multiple cids per msgid
This allows up to a configurable number (16 in this patch)
CIDs to be associated with a single messageid.  It still assumes
that CIDs will be unique per normalised subject, for the purpose
of finding matching CIDs in message.c, so it only stores one
"correct" CID per messageid.

NOTE: the cunit tests are not updated yet.  I'll come back and
do those after review, because it's not worth the work if this
design is going the wrong way.