Bug #2897 - fix check for duplicate service names
Things go horribly wrong if two separate SERVICES entries in cyrus.conf
have the same service name.  In particular, the accept() call that
allows service processes to receive a connection from a client are
locked using a fcntl lock on a filename (this isn't strictly necessary
on Linux but the code is there for other OSes).  That filename is built
using the name of the entry in SERVICES.  If two or more services have
the same name, and the prefork= option is used, one but one of them will
block in fcntl() instead of in accept(), and no client will be able to
connect to them.

When master reads cyrus.conf, it can and should check for this.  And in
fact it used to, there is code in there designed to detect duplicate
names, issue a complaint to syslog, and fail to start master.  This
could be better; for example, it might be capable of detecting a
*second* mistake in the file before exiting.  But at least it's clear.

Except...that check was broken in commit

  36d5924 fixed bug #2582 - friendly handling of removed services (via SIGHUP)

about 2 years after it was added in commit

  cb95f9f abort if we get two identical service names

which is by the same author.