Bug 3617: fix ACL check setting flags in delivery
When setting IMAP flags on a message during delivery, for example due to
a Sieve script requesting it, check for the 'p' = ACL_POST right instead
of the 'w' = ACL_WRITE or 't' = ACL_DELETEMSG right.  This seems to be
the correct interpretation of the following text in RFC2068

   [...] The currently defined standard rights are:
[...]
   w - write (STORE flags other than SEEN and DELETED)
[...]
   p - post (send mail to submission address for mailbox,
       not enforced by IMAP4 itself)

which implies that 'w' only applies for the IMAP STORE command.