Fix use after free during LMTP delivery
Found by Valgrind during the Annotator.set_user_flag_deliver test.

Invalid read of size 1
   at 0x4C292A4: __GI_strlen (mc_replace_strmem.c:284)
   by 0x5070B0A: make_key (duplicate.c:129)
   by 0x5070ED8: duplicate_mark (duplicate.c:224)
   by 0x408BD1: deliver_mailbox (lmtpd.c:604)
   by 0x40976C: deliver_local (lmtpd.c:793)
   by 0x409B93: deliver (lmtpd.c:884)
   by 0x40DA4F: lmtpmode (lmtpengine.c:1271)
   by 0x408012: service_main (lmtpd.c:327)
   by 0x411DE5: main (service.c:627)
 Address 0x7e3d200 is 0 bytes inside a block of size 37 free'd
   at 0x4C27D71: free (vg_replace_malloc.c:366)
   by 0x5085324: mailbox_read_header (mailbox.c:1216)
   by 0x5087281: mailbox_commit_header (mailbox.c:1929)
   by 0x508776B: mailbox_commit (mailbox.c:2040)
   by 0x506896C: append_commit (append.c:261)
   by 0x408B5A: deliver_mailbox (lmtpd.c:597)
   by 0x40976C: deliver_local (lmtpd.c:793)
   by 0x409B93: deliver (lmtpd.c:884)
   by 0x40DA4F: lmtpmode (lmtpengine.c:1271)
   by 0x408012: service_main (lmtpd.c:327)
   by 0x411DE5: main (service.c:627)

The duplicate key is constructed by pointing to the mailbox' uuid.
But that string is freed and reallocated during append_commit() if
a new user flag was added to the mailbox during the append.  That
may be unusual but it is what that test does.

This bug has been lurking for some time but was brought to the
surface when commit

9b3ff71 Bug #3690 - duplicate database - use uniqueid

changed over to always using the uuid for the duplicate DB.