Bug 3768: Makefile.am: Add -lsasl and -lssl as explicit dependecies
Makefile.am:
`libtool --config` reports on some systems for link_all_deplibs "yes" on
others "no".  To have a reliable build system, when linking the executables,
libsasl and libssl are now mentioned explictly as dependencies.  Otherwise
libtool with link_all_deplibs=no will not pass -lsasl and -lssl to the
linker and the it will fail.