mailbox: don't magically rebuild the cache
This is one of many misguided "magic repair" things that I put
into Cyrus 2.4.  I've never heard of it improving matters, but
it adds unpredictable behaviour.  Generally if the cache has
failed to exist, it's because something is broken that means
building a new cache won't work either.

Throw an IO error and make the sysadmin fix things, and chances
are you won't need to rebuild the cache anyway.