caldav: don't for message unlink
This is configurable for the admin already via the delayed_expunge
option - and keeping the files on disk allows for much better
undo/recovery options.

What might be worth considering is a mailbox option flag to turn on
immediate expunge (exposed via metadata like shared-seen) for
calendars with a high churn-rate of edited items.