Bug #3839: Cyrus::IMAP does not use -cafile/-capath.
When calling Cyrus::IMAP authenticate with -capath and/or -cafile it fails to
pass them to _starttls when starttls is used but not for authentication.

authenticate also fails to translate CApath and CAfile to -capath and -cafile,
so if CApath or CAfile are passed to authenticate, they won't be used.