[imap/imapd.c]: Allow RENAME to specify a partition, or a server, or both.
Again, the third parameter to a RENAME is more like a location than a
partition or server.

When a RENAME is issued, the command is proxied for MBTYPE_REMOTE
mailboxes.

Specify 's!p' to both transfer a mailbox on to the partition p of
server s.

Specify 'p' to move the mailbox between partitions on the current
backend.

Specify 's', where 's' is an invalid partition name on the current
backend server, to have the backend treat the argument as a server
address.