[sieve] fix backward compatibility for sieve index extension
There was a version of the bytecode that had the index extension but did
not update the bytecode codepoints, nor did it increment the bytecode
version number.  We use heuristics to determine if the index extension
support was included in the particular version of the bytecode.

This patch allows usage of these unnumbered versions of bytecode.