[sieve] bc_eval.c: sieve_eval_bc(): make ip point to first parameter
Before this change, when an action was switch'ed to, the ip might have
either pointed to the opcode in the bytecode, or to the previous
parameter, depending on whether the code was being switch'ed to
directly or was being reached as a result of fall-through.  This
change removes that ambiguity.

This also fixes a bug introduced by the imap4flags extension, where
older FILEINTO bytecode was not properly read.