Outlook 2013-compatible XLIST behaviour
Summary:
This patch changes XLIST behaviour so that it works the same as in LIST
in a special case. After this patch the output should be:
. LIST "" ""
* LIST (\Noselect) "/" ""
. OK Completed (0.000 secs)
. XLIST "" ""
* XLIST (\Noselect) "/" ""
. OK Completed (0.000 secs)

The meaningful component of the change is allowing an XLIST command with an
empty mailbox name argument to be used to query the top-level hierarchy
separator in the same way that LIST usually is (rfc 3501 s6.3.8).

I think this also changes Cyrus's LIST extended (rfc 5258) behaviour -- but
if so, the new behaviour is more correct.  It looks to me like Cyrus will
currently do the same thing with an extended LIST command when the mailbox
argument is empty, but rfc 5258 s3 specifically says it MUST NOT do this.

Test Plan:
Compiles, passes cunit and cassandane tests

Added LIST tests to cassandane

Reviewers: #imap_reviewers

Differential Revision: https://git.cyrus.foundation/D19