cyradm: hide value.priv and value.shared from user interface
Hide the imap protocol details "value.priv" and "value.shared" which
have been recently added to the cyradm info and mailboxconfig commands by
   [PATCH 2/3] cyradm: request and print value.shared and value.priv attribures
   [PATCH 3/3] cyradm: add support to set private annotations
from the user interface.

In the cyradm info command display it as "private" and "shared". In the
cyradm mboxconfig command replace it with an optional --private option
which switches from the default shared to the private version of an
annotation (e.g. "mboxconfig --private INBOX.Trash /specialuse \\Trash").

The new getmetadata and setmeatada commands also use "private" and "shared"
(though with a different display format).