Fix T156 by adding CAPAF_DASH_STUFFING flag into LMTP protocol description in deliver.c.
Messages delivered using 'deliver' to shared mailboxes were being rejected unless the shared mailbox
had the 'anonymous p' ACL set. 'deliver' was not providing 'AUTH =' extensions to 'MAIL FROM'
because the AUTH capability was not being detected. The AUTH capability was not being detected
because capabilities were being incorrectly parsed. Capabilities were being incorrectly parsed
because all except the last capability in LMTP is separated from the initial '250' by a hyphen
rather than a space. Adding the CAPAF_DASH_STUFFING flag causes hyphens to be considered as word
separators and the capability to be parsed correctly. This change makes the LMTP protocol
description in 'deliver.c' now match that in 'lmtpd.c'.