dav: normalise alive checks in GETFIELDS callers
If tombstones are an option, then we don't check for alive = 1
in the SQL, otherwise we do.

This means that in particular lookup_uid won't hit the version of
the record in another mailbox later (potentially) and return an error.

More uniqueness checks on UID will help too, but this is a start