The commit before removing all tabs
Welcome history investigator.

This is the very last commit in which in the tree has the mixed
tabs/spaces indent style.  It has been created to assist with
rebases.

This is a deliberately empty commit.  It will be given a tag
called "pretab" for easy reference.  A new commit with the
whitespace changes applied will be committed immediately after
this and called "posttab".

If you have a branch which you want to convert all the commits
into the new spacing style for easy rebase/cherry-pick, then
you can run:

   1) git checkout <mybranch>
   2) git rebase pretab
   3) git filter-branch --tree-filter tools/remove-tabs.pl pretab^..HEAD
   4) git rebase posttab

NOTE the "pretab^".  It is very important that you keep the ^ so that
this commit is the one that collects all the whitespace changes.

This will rewrite all the commits on your branch to the new whitespace
style and then rebase it on top of the posttab commit.  Since the
empty pretab commit will be rewritten to match posttab exactly by the
filter-branch process, the rebase will be clean and will skip the
pretab commit.

NOTE: you may need to make fixes at step 2 if your branch is old and
won't rebase cleanly on top of pretab.  Step 3 is noisy and slow, but
should always run correctly.