dlist: add DL_SFILE type and DLIST_SFILE flag
This is for representing files in situations in which a physical file
should not be created.

dl->sval contains the entire file contents (be careful about size!)
dl->nval is the file size
dl->oval is the offset within the protstream the dl was parsed from

dlist_parse and friends can be told to use DL_SFILE rather than
DL_FILE by setting the DLIST_SFILE bit in the flags parameter.