log the double a little earlier
This makes us slightly less likely to nuke the source mailbox from
the replica.  It's still not 100% safe, because something else could
have sync_logged on that mailbox due to an append or other change.

The only real way around this would be to do a lookup for the uniqueid
when a mailbox gets deleted (or keep a forward pointer in the mboxlist
entry) so that sync_client can detect the rename properly.