Avoid leaking idmap memory for erroneous JMAP POST requests.
Summary:
jmap_post allocates a hashmap for 1024 buckets. But if the HTTP
body could not be read or the client did not provide the mandatory
"application/json" HTTP header, this memory is not freed.

Test Plan:
This being my first patch, I am not sure what to make of this
section. All unit tests complete successfully, as do the cassandane tests.

Run Summary:    Type  Total    Ran Passed Failed Inactive
              suites     38     38    n/a      0        0
               tests    388    388    388      0        0
             asserts 828913 828913 828913      0      n/a

Reviewers: elliefm, #imap_reviewers

Reviewed By: elliefm, #imap_reviewers

Subscribers: elliefm

Differential Revision: https://git.cyrus.foundation/D70