Makefile.am: remove unnecessary conditionalisation
For conditionally-compiled binaries, libraries, etc, making their
inclusion in sbin_PROGRAMS et al conditional is adequate to ensure
they are not built when they are not desired.

It is safe to leave the specification of their respective _SOURCE,
_LDADD, et al as unconditional -- nothing will be built as a
direct consequence of these existing.

Further, making the _SOURCE specifications unconditional allows
automake to understand that these sources are always important and
need to be included in distribution tarballs (even if not configured
to build anything from them).  This means we don't need to juggle
them into EXTRA_DIST, either. :)

Note that we still do the `if FOO ... else EXTRA_DIST ... endif`
dance for manpages that document conditional features.  This is
because they should not be installed if their feature isn't, but
they must be included in the distribution tarball.