configure: somewhat improved handling of sqlite detection
separates the cyrusdb_sql-specific and general sqlite linkages
out into a single implementation, and then uses this.  this
fixes the problem where the only way to build/link the lib/sqldb
module was to --enable-http, even if you needed sqlite for
something else instead.

this is still not quite ideal: if you need sqlite for httpd or
objectstore only, and it's not in a default location (such that
you need --with-sqlite in order for it to be found), then suddenly
you've also got cyrusdb_sql enabled with sqlite as a backend

ideal future implementation:

 * --with-[mysql|pgsql|sqlite]  => to permit linkage of these sql engines
 * --enable-cyrusdb-sql         => to turn on cyrusdb_sql feature (needs
                                   at least one sql engine)
 * --enable-http                => to turn on http feature (needs sqlite)
 * --enable-objectstore         => to turn on objectstore feature (needs
                                   sqlite)