charset: export unfold and allow to skip whitespace
Define and export charset_unfold to unfold byte buffers according to
RFCs 2822, 2425 and 5545.

This required a change to the existing unfold routines, which always
emitted a whitespace character following a CRLF sequence. A caller
may now choose between emitting or omitting these whitespaces by
setting the CHARSET_UNFOLD_SKIPWS bit in flags.

For backward compatibily, the default is to emit whitespaces.