config2header: don't cast to union under SunCC
https://bugzilla.cyrusimap.org/show_bug.cgi?id=3918

Description From Marty Lee 2015-12-07 18:45:51 EDT (-) [reply]
Created an attachment (id=1563) [details]
Patch to config2header

The Sun C Compiler (Studio 12) is more strict than the GCC one, in that you
can't cast to a union - which happens a lot in imapopts.c

Patch attached to tweak the config2header file that creates imapopts.c, so that
the union casting is conditional on the compiler being used - i.e. if it's the
Sun Studio compilers, then don't cast to the union - otherwise do.

This is the same issue as found with 2.5.5, and the comments on that bug were
to not patch imapopts.c, but to patch the tool that generated it, so I have :-)

With this patch, cyrus 2.5.7 compiles cleanly on Solaris 10 (legacy system)
with Studio 12 compilers.