Allow replicated backends in a discrete Murder to both  actively participate in the Murder.
Summary:
With two backends (`imapb01` and `imapb02`) both configured to use a `servername: imapb0`,
the MUPDATE location used for mailboxes would consistently be `S imapb`.

The replication routines however will cause the second system (the replica) to issue a
copy of the MUPDATE commands (reservice, activate, deactivate, delete) for what is
otherwise to be expected is both the same mailbox, in the same location, as well as
with/for the same metadata to that mailbox.

This patch allows the MUPDATE server to suspend bailing out on the same command being
issued for the same mailbox in the same location a second time.

This implies that the backends' `servername` configuration option is something set to
the equivalent value between replication partners, and that the servername configured
resolves for the frontend. As such, one may use whichever failover methodology such as
a service IP address master negotiation between replication partners.

Effectively allows for hot-hot standby topologies in discrete murder topologies, or in
even more advanced topologies, with master-master replication, load-balanced, redundant
backends.

Test Plan:
  * Install three systems;

    # A frontend / mupdate master,
    # Two backends

  * Configure the backends as replicas of one another.

  * Ensure that the backends each have the exact same server name and partition layout.

  * Use the backends' `servername` in `serverlist` on the frontend.

  * Ensure that `servername` resolves (for the frontend) to at least one of the two
    backends.

  * Create a mailbox.

    Expect the replica to create the mailbox successfully as well.

  * Remove a mailbox.

    Expect the replica to remote the mailbox successfully as well.

Reviewers: #imap_reviewers

Differential Revision: https://git.cyrus.foundation/D84