Makefile.am: lex-fix workaround for debian flex bug
Debian flex 2.6.0-11 includes a patch series from upstream that
turns some previously yy_size_t variables back to int, for compatibility
with read(), but it's missed the last patch in the series, so yy_buf_size
remains a yy_size_t and triggers comparison-signedness warnings.

So: detect whether yy_n_chars is an int, and if so, make yy_buf_size one too.

Got sick of doing this by hand every build so in it goes :P