backup_mailbox_to_dlist: nuanced RECORDS inclusion
include RECORDS in dlist if the backup_mailbox has a records list,
even if the records list is empty.  do not include RECORDS in dlist
if the backup_mailbox has no records list.

use appropriate want_records value when retrieving backup_mailboxes
to control whether they contain a records list or not.