backupd: open backups in non-blocking mode
This is okay now that sync_client can comfortably handle an
IMAP_MAILBOX_LOCKED response.

What this means is that, if a locked backup is encountered during
a sync, instead of the whole sync stalling until the locked backup
becomes available, the sync can continue backing up other users,
and return to the locked one(s) later when they're hopefully no
longer locked.