perlmake.m4: work around MakeMaker uninstall crash bug
MakeMaker uninstall promises to show a deprecation warning and then
successfully do nothing.

Instead, it tries to read a file it never created, which thus
probably doesn't exist, in which case it dies.

Work around it by tweaking the generated Makefile to read the same
file it created.

This gets us closer to a working `make distcheck`, which I really
really really want.