com_err: make et_[foo]_error_table visible
This commit:

    6528d7a http_proxy.c, etc: HTTP/2 requests from a client can
            now be proxied to a backend (as HTTP/1.1) with the
            responses being sent back as HTTP/2

depends on the existence and visibility of et_http_error_table,
based on whatever version of com_err Ken has.  But the version of
com_err we include in the Cyrus distro didn't make this available.
Now it does.

Still to do probably: configure.ac checks for the existence of this
feature in whatever system-provided com_err we detect, and fall
back to using our version if it's not available.