message: use last occurrence of headers, not first
Where individual headers occur multiple times, DKIM signing
and verification prioritise occurrences in the reverse order
of which they appear in the header block (i.e. the oldest
headers, as long as new headers are prepended to the block).

The headers Cyrus cares about should generally only occur
once, so this change selects the same one to which DKIM
verification would have applied, if that occurred.

The exception to this is the Received header, which we parse
to extract the body->received_date, and of which we still use
the first (newest) occurrence.  (Note that if the header
X-DeliveredInternalDate exists, its last (oldest) value still
takes precedence over the one parsed out of Received.)

The full list of headers affected by this are:

    Bcc*
    Cc*
    Content-Description
    Content-Disposition
    Content-Id
    Content-Language*
    Content-Location
    Content-Md5
    Content-Transfer-Encoding
    Content-Type
    Date
    From*
    In-Reply-To
    Message-ID
    Reply-To*
    References
    Subject
    Sender*
    To*
    X-Me-Message-ID

(Items marked with a * would previously collate all values
from all occurrences of such headers, which was nice but not
strictly correct, and doesn't match DKIM behaviour.)