imapd.c: imapoptions: implement idle timeout
Use the value of the configuration variable "timeout" as an upper
limit in minutes for idle connections. To allow further
customization, add a new configuration option "imapidletimeout"
which, if greater than zero, will be used instead. The value
defaults to zero (not set).

RFC 2177 recommends that a client re-issue the IDLE command at
least every 29 minutes if it wishes to continue, otherwise the
server is free to treat the client as disconnected.

The rationale is that sometimes connections aren't properly
reset. Currently, a connection is not collected if it was in IDLE
state at that point. If this happens repeatedly, imapd keeps
accumulating dead connections which can cause DOS. This patch
solves the problem by forcing imapd to stop idling after
exceeding the configured timeout.