backup: honour the fulldirhash option when creating backup files
Previously, it acted as if fulldirhash was always set, whether it
was or not.  This may be a breaking change if you depended on this
behaviour in an environment with fulldirhash disabled.

This only affects the creation of new user backup files.  Backups
that already existed will remain functional in their original
location (because backups.db knows where they are), but new backups
will be created according to the fulldirhash option.