Preserve the unique ID of a mailbox.
A renamed mailbox doesn't really create a new mailbox, but more importantly,
a transferred mailbox doesn't really create a new mailbox -- it's just moved.

The use-case in which this would otherwise be problematic is in transferring
a mailbox between replicated backends. The replica master on the receiving
end would (prematurely) create a mailbox with a new uniqueid, provided the
command issued is a `LC1 LOCALCREATE user.john (PARTITION default)`.

This patch ensures that the command is, instead;

`LC1 LOCALCREATE user.john (PARTITION default UNIQUEID <uniqueid>)`.

It is not allowed for non-admins and non-proxyservers.