mailbox: store one modseq per call to dirty (aka, new modseq per UID changed in bulk changes)
The cost is negligable, because:

a) if we're not in conversations mode it's all NOOPs anyway
b) we already have a lock on the conversations DB, so the value won't change anyway,
   so the total extra cost is one more read of the counters file per mailbox change.
c) we have 64 bits to play with, extra numbers won't hurt any.

But this does guarantee that we can always calculate an ordered set of changes.  Every
single change has its own MODSEQ