Move the REID operation from cyr_info to mbtool
It's always struck me as odd that re-IDing a mailbox is a job for
cyr_info, rather than for mbtool.  From their respective man pages:

    *cyr_info* is a tool for getting information from Cyrus.
    *mbtool* is a tool for performing various actions on the indexes
		of a list of mailboxes.

Wouldn't the "reid" action then fall under the purview of mbtool?
It looked to me as though it should be pretty easy to move it, so I've
tried just that.