Option to build Cyrus statically.
This patch provides configure option to enable static builds. (Actually
it just uses the `--enable-static` and `--enable-shared` options
provided by `autoconf`). When configured statically, we ensure that we
link with `libstdc++` which is a dependency since, we link with
`libxapian`.

What does it actually mean to build Cyrus statically?
Only the binaries and libraries we build(`libcyrus_*) as a part of
`cyus-imapd` are built and linked statically. Only static archives(`.a`)
are built. All the external libraries are still dynamically linked.