mailbox: only assert the modseq status AFTER updating it if not silent
In theory the other way was fine as well, because we should always be
having at least the modseq we just read - but with annotation updates
we read and rewrite the record from disk without updating our working
copy.  Since only the modseq gets changed with an annotation rewrite,
it's safe to write our modifications down so long as we don't insist
that the modseq was higher FIRST.

This commit should be reverted once we can do both annotation and flag
updates on a msgrecord with just a single rewrite_index_record.