Revert "Makefile.am: don't bundle generated perl Makefile.PLs either"
This reverts commit c07164fa7cc7e80dbc0d0c32bd64ebc5b1c38f5a.

This would be a good commit, but it causes "make distcheck" to fail
late in the game because the perl-specific Makefiles don't know how
to rebuild Makefile.PL from Makefile.PL.in (because the top level
configure does that), and so they bail out on what looks like a
safety check at "make distclean".

I really don't feel like fighting with conflicting build system
requirements right now, so just reverting this.