[Idle] Enssure idled is working as expected.
By setting the timeout to 60 seconds(which is the timeout if
`imapidlepoll` isn't set in the `imapd.conf` and ensure that `idled`
responds immediately, we make sure that `idled` is working as expected.
This test fails if the response takes longer than 5 seconds, it is a
problem when running on slow machines. Will bump the 5 second time based
on how the test works on CI.