[Bug 2086] Should be able to set special value twice.
Currently when setting the special value of a folder twice to same
value, the `setmetadata` command is rejected. This patch checks if the
mailbox we are trying to set the special use flag on is the same
folder(with the same flag), we don't error, and return a success
instead.