[Docs] Symlinked test programs get sphinx included docs.
Move the core information into an asset.
Do a sphinx .. include:: to pull in that information into each
man page with a header to note that the command itself is a
symlink.

This means if the underlying imtest changes, all the symlinked
program docs will also stay up to date.