user: fix sieve script path for users with dots in their names
user_sieve_path() and user_renamesieve() in imap/user.c also need to
take into account that user "foo.bar" with a sieve directory of
/var/spool/sieve has his scripts in a /var/spool/sieve/f/foo^bar/ directory
and not in /var/spool/sieve/f/foo.bar/.

Previously, renaming an user account from or to a name with a dot would
either fail to rename that user sieve script directory or would rename it
to a wrong name (with a dot instead of '^').

Fix this by adding dot-translating code from imap/lmtp_sieve.c also to
imap/user.c, just as we did for timsieved earlier.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>