[build] Workarounds to work with makemaker.
`make distcheck` fails in the Debian testing because of this[1] change in
`ExtUtils::MakeMaker`.

To workaround that problem, this commit adds the `MANIFEST` files to the
build src directories using `AC_CONFIG_FILES`. I tried with
`AC_CONFIG_LINKS` but since the resulting symbolic links point to read
only files, that didn't work. So using `AC_CONFIG_FILES` instead.

Using `AC_CONFIG_FILES` results in a couple of warnings from automake
saying there is a circular dependency issues and that rule isn't being
ignored.

At this point there are too many workarounds in place for autotools to
work with the perl modules in Cyrus. Perhaps it is time to rethink how
we work with the perl modules.

[1] -
https://github.com/Perl-Toolchain-Gang/ExtUtils-MakeMaker/commit/5742aeb93564a24634adc6055aa2a745d228d287