sync_support: downgrade recoverable SYNCERRORs to SYNCNOTICE
Assuming that SYNCERROR means what I think it (should mean), namely "I
can't possible continue and will need an operator to intervene", all the
cases that are recoverable by doing a full resync should only warn.

The only thing I'm not sure about is if a full resync can ever fail for
one of these reasons. If it can, then there's now nothing to alert the
operator, and this patch probably should be applied as-is (at the cost
of someone getting woken up :/ )