caldav: allow zero length events
Even though the standard says they are not legit, in practice they happen
lots, particularly for all-day events, and people generate them.

Our choices are to strip DTEND==DTSTART or DURATION==P0D to be exact with
the standard, or to just accept them and not care.  I'm pretty happy with
accept and don't care, it's a crappy restriction in the standard.

                  ; Either 'dtend' or 'duration' MAY appear in
                  ; a 'eventprop', but 'dtend' and 'duration'
                  ; MUST NOT occur in the same 'eventprop'.
                  ;
                  dtend / duration /

So a zero length event just has neither.  Since there are already multiple
ways of writing any positive duration, it doesn't hurt to allow multiple ways
of writing a zero duration.