Try to clarify documentation in re: Partitions
This is a first attempt to clear up some confusion in regards to the
nameing of partitions and the use of the various "partition-name"
configuration directives in imapd.conf(5).

It's my goal to eliminate all use of common words like "default" or
"archive" in the names of partitions in samples and descriptive
language, lest people perceive such use to mean that those names
*must* be used.

For example, rather than using the following:
    defaultpartition: default
    partition-default: /var/spool/cyrus

we can use:
    defaultpartition: part1
    partition-part1: /var/spool/cyrus

The file /lib/imapoptions has been lightly modified in service to this
cause, and to correct typos in the definition of searchpartition-name
settings.