jmap: remove experimental JMAP authentication support
This patch removes the JMAP authentication database, HTTP
handlers and tools. The JMAP specification draft deprecated
JMAP authentication and there is no reason to implement a
custom authentication mechanism in Cyrus if clients are
not going to support it.

The JMAP settings for an authenticated account are now returned
for GET requests on the /jmap/ base URL. Clients can follow HTTP
redirects from the /.well-known/jmap URL to get there.