Clarify documentation regarding default sync_log
The Cyrus IMAP replication subsystem assumes a default sync_log
channel named "" (the two-character string U+22 U+22).  If one modifies
the sync_log_channels: directive with a name for another sync_log
channel, such as "squatter" then the default "" channel vanishes, unless
it is also specified explicitly. [issue #2248]

This commit is a first draft at making this clear.  Changes are made to
both the manpage for squatter(8) and to the Admin Guide, Replication
pages.

I feel that we actually need to produce more generalized documentation
of the sync_log mechanism, and then link to that from all points
which deal with sync_log usage, as these will no doubt grow due to the
usefullness of the sync_log functionality.

Addresses #2248.