sync_log: new log record "RENAME old new" and sync_log_rename()
If the mailbox ownership changes across the rename, both mailboxes
will be batched with both users, so it doesn't matter which user
gets batched first.

For backward compatibility, sync_log_rename produces:

    RENAME a b
    MAILBOX a
    MAILBOX b

in the sync log.  This will avoid breaking cases where a replication
log is being reprocessed by an older build that doesn't understand
RENAME records.  The unrecognised RENAME will be ignored, and the
rename will be replicated in the traditional style by the MAILBOX
double.