Convert and move doc/internal
This commit involves converting all of the documentation in the old
/doc/internal location into RestructuredText, and puts it in various
branches of /docsrc/imap/developer:

	guidance - Documents which explain how or why things are done
	thoughts - Random notes and ideas
	API - Application Programming Interfaces

Many of the source documentation is very old, and no real effort
has been made to ensure that it is up to date or accurate.

Also, conf.py was updated to bring the copyright up to 2018.