Return an exit status even if we fork using a filehandle
In some cases, run_command() would fork using a pipe:

        # Use the fork()ing form of open()
        my $pid = open $fh,'|-';
        die "Cannot fork: $!"
            if !defined $pid;
        if ($pid)
        {
            # parent process
            $self->_add_child($binary, $pid, $options->{handlers}, $fh);
            print $fh ${$data};
            close ($fh);
            return $pid;
        }

The problem is, the close ($fh) there waits for the child to exit,
which means later in reap_command(), waitpid($pid, 0) returns -1
since there are no children, and our caller gets back undef.

This fixes that by returning the child status if we know we already
have one and bypassing reap_command() entirely.