Split the system_flags field in struct index_record.
The `system_flags` field in `struct index_record` is used to store the
following flags on disk:

-- System flags:
 FLAG_ANSWERED
 FLAG_FLAGGED
 FLAG_DELETED
 FLAG_DRAFT
 FLAG_SEEN

-- Internal flags for bookeeping:
 FLAG_INTERNAL_SPLITCONVERSATION
 FLAG_INTERNAL_NEEDS_CLEANUP
 FLAG_INTERNAL_ARCHIVED
 FLAG_INTERNAL_UNLINKED
 FLAG_INTERNAL_EXPUNGED

This patch, removes the use of `system_flags` to store the internal
flags in memory. Instead, we now have an `internal_flags` field in `struct
index_record` representing the internal flags.

The on-disk representation continues to be the same.

The methods `mailbox_buf_to_index_record()` and
`mailbox_index_record_to_buf()` serialise and deserialise these fields.