TestCase: reimplement :needs_foo_bar as a filter instead
Test::Unit::TestCase::list_unit() inverts the usual inheritance chain
in a really annoying way, so if you override it you basically can't
call SUPER to extend the default behaviour, uggh.  Just because you
can, doesn't mean you should!

Implementing it as a filter is less than ideal, as it confuses the
number-of-tests counter, but it's what we've got to work with.