cyr_buildinfo: zero-deps backport from master
The version on master uses libjansson to build the report (and
supports some additional formatting features).

This is a crude implementation that does the string munging itself,
and thus does not alter 2.5's dependencies.

Many of the features it reports as "false" are not features that
actually exist in this version of Cyrus, so don't expect them to
be configurable to true :)