sieve: HORRIBLE EVIL BAD NO GOOD LEAK WORKAROUND
This is really awful, but basically the "atleast" mechanism is horrible
and someone somewhere forgot one, and we write unallocated memory.

This just makes sure we have a big fat slop so we never run out.  It's
disgusting, but better than crashing.

We're gonna toss this whole thing away and replace it with something
better one day.  Until then, this nasty workaround thing remains.