intermediates: edge trigger updates after any actions
More able to be correct this way, at slightly higher cost :)

Intermediates still suck, but this will hopefully make them suck less.

Known broken edge case (still!) when a mailbox gets a new uniqueid,
there isn't currently a tombstone for the old uniqueid which can be
used by JMAP to issue a "removed" for it.  We need to generate an
intermediate record in the DELETED namespace for that if we want to
be perfect.